Revert "meta: Switch from pnpm to bun"
This would have been cool but jsr does not work on NixOS with seemingly
no way to manually compile the jsr binary that jsr-npm automatically
downloads.

Although it may be possible to use jsr through deno (which hasn't been
updated in nixpkgs yet), it's *much* simpler to simply use the existing
node workflow.