docs: Add stretch goals
I wrote this before realizing that it's possible to make my own Tailwind
CSS plugin and use that. This could help with simplifying the build
process in such a way that the plugin can parse the YAML files directly
without having to run ts-node to generate an output CSS file first.

By removing the CSS output, it should become theoretically possible to
publish the TypeScript-only plugin on JSR as well.