refactor: Merge Base16Scheme with Base16Yaml
This is a trivial change that simplifies the code base by relying on the
fact that we don't care what happens to the initial yaml structure when
passing it into the class.

Further simplification can be done to remove unnecessary values that are
no longer needed from the initial yaml structure.

This change also has the advantage of showing a more useful error
message to the user as needed.