chore: Make tests work again
Also removed an unused `return this` and fixed the name type assuming
base16- at the end for the inherited yaml.

The text-100 and bg-800 classes have been removed from the tests since I
currently don't change the user's tailwind config with the plugin at the
same time, although it should be possible.