meta: Rename Base16Yaml to Base16Scheme
This makes the scheme class generic and able to be used in a wider
variety of contexts. Indeed, we can support any file format as long as
there's a way to turn it into a JavaScript object. End-users should also
be able to create their own Base16Schemes if this class is exported.

Note that the file wasn't renamed from Base16Yaml to Base16Scheme in
this commit to avoid Git recognizing the changes as a new file.