refactor(palette): Store colors in array; remove isValid
We don't actually care about which baseXX value a color is since the
colors will always be in order, so we can use an array instead.

This should make it easier to support other baseXX schemes in the future
if desired, although base16 is already a lot of colors to choose from.

Since we're already using a constructor that fails on errors, we can get
rid of the isValid checks and simplify to just creating a new object
instead. This also helps us get rid of the duplicate for loop
iterations.